Add shops and npcs in to the api. Remove iseason. API not standalone atm. Will worry about that later.

+++ src/main/java/joshie/harvestmoon/shops/Purchaseable.java
@@ -1,15 +1,15 @@
 package joshie.harvestmoon.shops;
 
-import joshie.harvestmoon.api.core.IPurchaseable;
+import joshie.harvestmoon.api.shops.IPurchaseable;
 import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.item.ItemStack;
 import net.minecraft.world.World;
 
 public class Purchaseable implements IPurchaseable {
     private ItemStack[] stacks;
-    private int cost;
+    private long cost;
     
-    public Purchaseable(int cost, ItemStack... stacks) {
+    public Purchaseable(long cost, ItemStack... stacks) {
         this.cost = cost;
         this.stacks = stacks;
     }
